refactor fungible
