+++ x/fungible/keeper/msg_server_udpate_zrc20_liquidity_cap_test.go
@@ -3,6 +3,8 @@ package keeper_test
 import (
 	"testing"
 
+	authoritytypes "github.com/zeta-chain/zetacore/x/authority/types"
+
 	"cosmossdk.io/math"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/stretchr/testify/require"
@@ -10,12 +12,14 @@ import (
 	"github.com/zeta-chain/zetacore/testutil/sample"
 	"github.com/zeta-chain/zetacore/x/fungible/keeper"
 	"github.com/zeta-chain/zetacore/x/fungible/types"
-	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
 
 func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 	t.Run("can update the liquidity cap of zrc20", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
 		admin := sample.AccAddress()
 		coinAddress := sample.EthAddress().String()
@@ -24,7 +28,8 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 		foreignCoin.LiquidityCap = math.Uint{}
 		k.SetForeignCoins(ctx, foreignCoin)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group2)
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		// can update liquidity cap
 		_, err := msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
@@ -38,6 +43,8 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 		require.True(t, found)
 		require.True(t, coin.LiquidityCap.Equal(math.NewUint(42)), "invalid liquidity cap", coin.LiquidityCap.String())
 
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
+
 		// can update liquidity cap again
 		_, err = msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
 			admin,
@@ -50,6 +57,8 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 		require.True(t, found)
 		require.True(t, coin.LiquidityCap.Equal(math.NewUint(4200000)), "invalid liquidity cap", coin.LiquidityCap.String())
 
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
+
 		// can set liquidity cap to 0
 		_, err = msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
 			admin,
@@ -61,6 +70,8 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 		coin, found = k.GetForeignCoins(ctx, coinAddress)
 		require.True(t, found)
 		require.True(t, coin.LiquidityCap.Equal(math.ZeroUint()), "invalid liquidity cap", coin.LiquidityCap.String())
+
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		// can set liquidity cap to nil
 		_, err = msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
@@ -76,7 +87,10 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 	})
 
 	t.Run("should fail if not admin", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
 		admin := sample.AccAddress()
 		coinAddress := sample.EthAddress().String()
@@ -85,10 +99,11 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 		foreignCoin.LiquidityCap = math.Uint{}
 		k.SetForeignCoins(ctx, foreignCoin)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, false)
 
 		_, err := msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
-			sample.AccAddress(),
+			admin,
 			coinAddress,
 			math.NewUint(42),
 		))
@@ -97,12 +112,16 @@ func TestMsgServer_UpdateZRC20LiquidityCap(t *testing.T) {
 	})
 
 	t.Run("should fail if zrc20 does not exist", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
 		admin := sample.AccAddress()
 		coinAddress := sample.EthAddress().String()
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group2)
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		_, err := msgServer.UpdateZRC20LiquidityCap(ctx, types.NewMsgUpdateZRC20LiquidityCap(
 			admin,
