+++ x/fungible/keeper/msg_server_update_zrc20_paused_status_test.go
@@ -7,16 +7,20 @@ import (
 	"github.com/stretchr/testify/require"
 	keepertest "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/testutil/sample"
+	authoritytypes "github.com/zeta-chain/zetacore/x/authority/types"
 	"github.com/zeta-chain/zetacore/x/fungible/keeper"
 	"github.com/zeta-chain/zetacore/x/fungible/types"
-	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
 
 func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 	t.Run("can update the paused status of zrc20", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
 		admin := sample.AccAddress()
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
 
 		assertUnpaused := func(zrc20 string) {
 			fc, found := k.GetForeignCoins(ctx, zrc20)
@@ -38,7 +42,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertUnpaused(zrc20B)
 		assertUnpaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupEmergency, true)
 
 		// can pause zrc20
 		_, err := msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -54,7 +58,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertPaused(zrc20B)
 		assertUnpaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group2)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		// can unpause zrc20
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -69,7 +73,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertPaused(zrc20B)
 		assertUnpaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupEmergency, true)
 
 		// can pause already paused zrc20
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -84,7 +88,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertPaused(zrc20B)
 		assertUnpaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group2)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		// can unpause already unpaused zrc20
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -99,7 +103,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertPaused(zrc20B)
 		assertUnpaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupEmergency, true)
 
 		// can pause all zrc20
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -116,7 +120,7 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 		assertPaused(zrc20B)
 		assertPaused(zrc20C)
 
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group2)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, true)
 
 		// can unpause all zrc20
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
@@ -135,10 +139,13 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 	})
 
 	t.Run("should fail if invalid message", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
+
 		admin := sample.AccAddress()
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
 
 		invalidMsg := types.NewMsgUpdateZRC20PausedStatus(admin, []string{}, types.UpdatePausedStatusAction_PAUSE)
 		require.ErrorIs(t, invalidMsg.ValidateBasic(), sdkerrors.ErrInvalidRequest)
@@ -148,18 +155,24 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 	})
 
 	t.Run("should fail if not authorized", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
 
+		admin := sample.AccAddress()
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupEmergency, false)
+
 		_, err := msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
-			sample.AccAddress(),
+			admin,
 			[]string{sample.EthAddress().String()},
 			types.UpdatePausedStatusAction_PAUSE,
 		))
 		require.ErrorIs(t, err, sdkerrors.ErrUnauthorized)
 
-		admin := sample.AccAddress()
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupAdmin, false)
 
 		_, err = msgServer.UpdateZRC20PausedStatus(ctx, types.NewMsgUpdateZRC20PausedStatus(
 			admin,
@@ -171,10 +184,15 @@ func TestKeeper_UpdateZRC20PausedStatus(t *testing.T) {
 	})
 
 	t.Run("should fail if zrc20 does not exist", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.FungibleKeeper(t)
+		k, ctx, _, _ := keepertest.FungibleKeeperWithMocks(t, keepertest.FungibleMockOptions{
+			UseAuthorityMock: true,
+		})
+
 		msgServer := keeper.NewMsgServerImpl(*k)
+
 		admin := sample.AccAddress()
-		setAdminPolicies(ctx, zk, admin, observertypes.Policy_Type_group1)
+		authorityMock := keepertest.GetFungibleAuthorityMock(t, k)
+		keepertest.MockIsAuthorized(&authorityMock.Mock, admin, authoritytypes.PolicyType_groupEmergency, true)
 
 		zrc20A, zrc20B := sample.EthAddress().String(), sample.EthAddress().String()
 		k.SetForeignCoins(ctx, sample.ForeignCoins(t, zrc20A))
