+++ x/observer/keeper/msg_server_update_keygen.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"context"
+
 	authoritytypes "github.com/zeta-chain/zetacore/x/authority/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
