+++ x/observer/keeper/msg_server_update_observer.go
@@ -3,6 +3,7 @@ package keeper
 import (
 	"context"
 	"fmt"
+
 	authoritytypes "github.com/zeta-chain/zetacore/x/authority/types"
 
 	errorsmod "cosmossdk.io/errors"
